remove obsolete CONFIG var
