+++ api/src/routers/middlewares.ts
@@ -1,9 +1,9 @@
 import { Validator } from 'express-json-validator-middleware';
-import { CONFIG } from '../config';
 import { hasValidApiKey } from '../middlewares/api-key';
 import { isAuth } from '../middlewares/authentication';
+import { ConfigurationService } from '../services/configuration-service';
 
-const { serverSecret, apiKey } = CONFIG;
+const { serverSecret, apiKey } = ConfigurationService.getInstance().config;
 
 export const validator = new Validator({ allErrors: true });
 export const validate = validator.validate;
