+++ api/src/routes/server-info/index.ts
@@ -1,6 +1,5 @@
 import { NextFunction, Request, Response } from 'express';
 import { StatusCodes } from 'http-status-codes';
-import { CONFIG } from '../../config';
 import { ILogger } from '../../utils/logger';
 import * as os from 'os';
 import { ConfigurationService } from '../../services/configuration-service';
@@ -10,10 +9,11 @@ export class ServerInfoRoutes {
 
 	getServerInfo = async (req: Request, res: Response, next: NextFunction): Promise<void> => {
 		try {
+			const config = this.configService.config;
 			const hostname = os.hostname();
-			const commitHash = CONFIG.commitHash || 'not defined';
+			const commitHash = config.commitHash || 'not defined';
 			const identityId = this.configService.serverIdentityId || 'not defined';
-			const version = CONFIG.apiVersion || 'not defined';
+			const version = config.apiVersion || 'not defined';
 
 			res.status(StatusCodes.OK).send({
 				commitHash,
