+++ api/src/services/configuration-service.ts
@@ -52,6 +52,7 @@ export class ConfigurationService {
 
 	constructor() {
 		this.logger = Logger.getInstance();
+		this.assertConfig(this.config);
 	}
 
 	public static getInstance(): ConfigurationService {
@@ -59,6 +60,29 @@ export class ConfigurationService {
 			ConfigurationService.instance = new ConfigurationService();
 		}
 		return ConfigurationService.instance;
+	}
+
+	assertConfig(config: Config) {
+		if (config.serverSecret === '<server-secret>') {
+			console.error('please replace the default values!');
+		}
+
+		if (config.serverSecret.length !== 32) {
+			throw Error('Server secret must to have a length of 32!');
+		}
+
+		// apiKey can be empty if the host decides so
+		// commitHash is set automatically during deployment
+		const optionalEnvVariables = ['apiKey', 'commitHash'];
+		Object.values(config).map((value, i) => {
+			if (isEmpty(value) && (isNaN(value) || value == null || value === '')) {
+				if (optionalEnvVariables.includes(Object.keys(config)[i])) {
+					return;
+				}
+
+				console.error(`env var is missing or invalid: ${Object.keys(config)[i]}`);
+			}
+		});
 	}
 
 	// Ensure that on the db there is the declared root identity
